+++ react.js
@@ -2,26 +2,30 @@ const React = require('react')
 const reactTestingLibrary = require('react-testing-library')
 const { IntlProvider } = require('react-intl')
 const { MockedProvider } = require('react-apollo/test-utils')
+const { path, insertAll, reject, isNil, find } = require('ramda')
 const paths = require('./modules/paths')
 
 const pkg = require(paths.resolveAppPath('package.json'))
 
-let defaultLocale = 'en-US'
-if (pkg.vtexTestTools && pkg.vtexTestTools.defaultLocale) {
-  defaultLocale = pkg.vtexTestTools.defaultLocale
+const getLocale = (optionsLocale) => {
+  const defaultLocales = ['en', 'en-US']
+  const pkgLocale = path(['vtexTestTools', 'defaultLocale'], pkg)
+
+  const locales = reject(isNil, insertAll(optionsLocale, pkgLocale, defaultLocales, []))
+  const localeExists = locale => paths.pathExists(`../messages/${locale}.json`)
+
+  return find(localeExists, locales)
 }
 
-const defaultMessages = require(paths.resolveAppPath('../messages/' + defaultLocale + '.json'))
-
 const customRender = (node, options = {}) => {
-  const intlProps = {
-    locale: defaultLocale,
-    messages: defaultMessages,
-  }
+  const locale = getLocale(path(['locale'], options))
+  const messages = locale 
+    ? require(paths.resolveAppPath(`../messages/${locale}.json`)) 
+    : {}
 
-  if (options.locale) {
-    intlProps.locale = options.locale
-    intlProps.messages = require(paths.resolveAppPath('../messages/' + options.locale + '.json'))
+  const intlProps = {
+    locale: locale,
+    messages: messages,
   }
 
   const apolloProps = options.graphql